Add a CXX flag that the code said it needed
When I compiled before on MSVC, the warning was as follows:

"warning C4530: C++ exception handler used, but unwind semantics are
not enabled. Specify /EHsc"

That sounds reasonable, but it seems strange to me that this isn't
handled by the flatbuffers CMakeLists.txt. Also, I don't see this
documented at all in the flatbuffers repo or used in their CI. I don't
generally work on windows, so I don't know if this is just something
that every windows dev knows.